+++ resources/views/categories/create.blade.php
@@ -1,11 +1,17 @@
-{!! form_open([
-    'id'     => 'edit-form',
-    'role'   => 'form',
-    'method' => 'POST',
-]) !!}
+{!! $this->makePartial('paolorox.restapro::partials/settings_tabs') !!}
+
+<div class="row-fluid">
+    {!! form_open(current_url(),
+        [
+            'id' => 'form-widget',
+            'role' => 'form',
+            'method' => 'PATCH',
+        ]
+    ) !!}
 
     {!! $this->renderFormToolbar() !!}
 
     {!! $this->renderForm([], true) !!}
 
 {!! form_close() !!}
+</div>