+++ resources/views/categories/index.blade.php
@@ -1,3 +1,5 @@
+{!! $this->makePartial('paolorox.restapro::partials/settings_tabs') !!}
+
 {!! $this->renderListToolbar() !!}
 
 {!! $this->renderListFilter() !!}