+++ src/Http/Controllers/Units.php
@@ -46,6 +46,6 @@ class Units extends AdminController
     {
         parent::__construct();
 
-        AdminMenu::setContext('units', 'production');
+        AdminMenu::setContext('settings', 'system');
     }
 }